Restore history playback from local storage, impliment last known position feature
