+++ map_service/models.py
@@ -1,7 +1,7 @@
-from mongoengine import Document, EmbeddedDocument, fields
+from mongoengine import Document, fields
 
 
 class LkState(Document):
     o_id = fields.StringField(required=True)
-    lk_state = fields.StringField(required=True)
+    lk_properties = fields.DynamicField(required=True)
     lk_geo_json = fields.PointField()
