+++ map_system/static/map_system/local/application_options.js
@@ -11,7 +11,7 @@ var ApplicationOptions = {
             header: 'grey'
         }
     },
-    constance:{
+    constance: {
         WEB_SOCKET_URL: "ws://geo.knnect.com:9080/ws/data",
         WEB_SOCKET_SERVER: 'aws4.knnect.com', /*TODO: Deprecated */
         WEB_SOCKET_PORT: 9764, /*TODO: Deprecated */
@@ -27,10 +27,8 @@ var ApplicationOptions = {
         NOTIFY_WARNING_TIMEOUT: 3000,
         NOTIFY_DANGER_TIMEOUT: 9000
     },
-    messages:{
-        app:{
-
-        }
+    messages: {
+        app: {}
     },
     leaflet: {
         iconUrls: {
@@ -46,15 +44,15 @@ var ApplicationOptions = {
     locale: {
         type: 'sin',
         sin: {
-            websocket:{
-                errors:{
+            websocket: {
+                errors: {
                     'connection': 'වෙබ් සොකට් සම්බන්ධතාවය විසන්ඳිවී ඇත කරුණාකර ඔබගේ අන්තර්ජාල සබදතාවය පරීක්ශාකර බලන්න  '
                 }
             }
         },
         eng: {
             websockt: {
-                errors:{
+                errors: {
                     'connection': 'Something went wrong when trying to connect to WebSocket </br> <b>Please check your internet or Network connection<b/>'
                 }
             }
