+++ map_system/templates/map_system/content/nav_bar.html
@@ -27,7 +27,7 @@
             </li>
 
             <li class="dropdown">
-                <a href="#" class="dropdown-toggle" data-toggle="dropdown"><span style="color: #f9fdff;">Signed-in as: <%= user.username + "@" + user.domain  %></span></a>
+                <a href="#" class="dropdown-toggle" data-toggle="dropdown">More</a>
                 <ul class="dropdown-menu" role="menu">
                     <li>
                         <a href="#" data-toggle="collapse" data-target=".navbar-collapse.in"
